Updates to the code for generation

+++ _Production/Scripts/Data/TerrainGenerationData.cs
@@ -1,3 +1,4 @@
+using System.ComponentModel.DataAnnotations;
 using Godot;
 
 namespace FT.Data;
@@ -15,7 +16,7 @@ public partial class TerrainGenerationData : Resource
     
     [ExportCategory("Terrain Data")]
     [Export] public int Seed { get; private set; } = 123456789;
-    [Export] public ushort CellSize { get; private set; } = 5;
-    [Export] public ushort Rows { get; private set; } = 20;
-    [Export] public ushort Cols { get; private set; } = 20;
+    [Export] public byte CellSize { get; private set; } = 5;
+    [Export] public byte Rows { get; private set; } = 20;
+    [Export] public byte Cols { get; private set; } = 20;
 }
